Improved Password Field
When entering the correct password, it know disables the lock screen and shows the proper homepage. This is a quality of life improvement so there is actually a home screen. Keep in mind that this password system is still very basic and should not be used in a production environment. There is still also scalability issues in the project.